Debug @70543ccbb4bfdaf813fb4e5ed39ce592995cc922.

+++ app/models/division_display_sc_item.rb
@@ -13,7 +13,7 @@ class DivisionDisplayScItem < ActiveRecord::Base
   include PaperTrailable
 
   before_destroy do
-    self.featured_content_selection.destroy
+    self.featured_content_selection.destroy if self.featured_content_selection
   end
 
 end
